Bug fixes for HYSPLIT tools
Corrected a bug in which the tdump2nc would sometimes incorrectly
label the variable dimensions in the netcdf file.

Fixed a bug in which trajectories starting in the first or last hour
of the day were not bracketed in time by meteorological data. For
trajectories starting during these hours, the prior or next day's met
data is used in the CONTROL file.

Signed-off-by: Christopher Holmes <cdholmes@fsu.edu>